+++ src/main/java/com/inti/controller/ChauffeurController.java
@@ -6,15 +6,18 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.inti.entities.AvisClient;
 import com.inti.entities.Chauffeur;
 import com.inti.entities.Planning;
+import com.inti.entities.Vehicule;
 import com.inti.service.interfaces.IAvisClientService;
 import com.inti.service.interfaces.IChauffeurService;
 import com.inti.service.interfaces.IPlanningService;
 import com.inti.service.interfaces.IReservationService;
+import com.inti.service.interfaces.IStatistiqueService;
 
 @RestController
 @CrossOrigin
@@ -31,6 +34,9 @@ public class ChauffeurController {
 	
 	@Autowired
 	IReservationService reservationService;
+	
+	@Autowired
+	IStatistiqueService statistiqueService;
 
 	// Affiche tous les chauffeurs
 	@GetMapping("chauffeur")
@@ -57,6 +63,11 @@ public class ChauffeurController {
 	@GetMapping("chauffeur/avisClient/{idChauffeur}")
 	public List<AvisClient> affichageAvisClient(@PathVariable("idChauffeur") Long idChauffeur) {
 		return avisClientService.findByChauffeur(idChauffeur);
+	}
+
+	@GetMapping("statistique/chauffeurBest")
+	public List<Chauffeur> chauffeursOrderBy(@RequestParam(required = false) String type){
+		return chauffeurService.chauffeursOrderBy(type);
 	}
 
 }
