+++ src/main/java/com/inti/controller/StatistiqueController.java
@@ -5,6 +5,7 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.inti.entities.Administrateur;
@@ -19,14 +20,8 @@ public class StatistiqueController {
 	@Autowired
 	IStatistiqueService statistiqueService;
 	
-	@GetMapping("statistique/chauffeurBest")
-	public List<Chauffeur> chauffeursOrderBy(){
-		return statistiqueService.chauffeursOrderBy();
-	}
+
 	
-	@GetMapping("statistique/vehiculeConsoPlus")
-	public Vehicule vehiculeConsoPlus() {
-		return statistiqueService.vehiculeConsoPlus();
-	}
+	/**/
 	
 }
