+++ src/main/java/com/inti/entities/Utilisateur.java
@@ -110,4 +110,10 @@ public abstract class Utilisateur implements Serializable {
 		this.adresse = adresse;
 	}
 
+	@Override
+	public String toString() {
+		return "Utilisateur [nom=" + nom + ", prenom=" + prenom + ", username=" + username + ", password=" + password
+				+ ", email=" + email + "]";
+	}
+
 }
