+++ src/main/java/com/inti/service/interfaces/IVehiculeService.java
@@ -14,5 +14,8 @@ public interface IVehiculeService {
 
 	void delete(Long idVehicule);
 	
+	Vehicule vehiculeConsoPlus();
+	
+	
 	List<Vehicule> findByAgence(Agence agence);
 }
