+++ src/main/resources/application.properties
@@ -5,9 +5,9 @@ server.port=7070
 spring.datasource.url=jdbc:mysql://localhost:3306/gestion_taxi_driver_db?createDatabaseIfNotExist=true
 spring.datasource.username=root
 
-spring.datasource.password=codeMySQL123
+spring.datasource.password=Paquito26
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 
 #hibernate properties
-spring.jpa.hibernate.ddl-auto=create-drop
+spring.jpa.hibernate.ddl-auto=update
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL5Dialect
